add detail/parsing/output_adapters.hpp
